Don't inspect XEvent queue in KUserTimestamp::updateUserTimestamp
updateUserTimestamp() used to fetch the current time by creating a
property notify event and wait till it is received by inspecting the
XEvent queue. That doesn't work with XCB anymore, but Qt provides
this functionality for us, so just use that.

At the same time the argument is changed from int to unsigned long
as that's what the method in QX11Info returns.

REVIEW: 112684

+++ staging/kinterprocesswindowing/src/kusertimestamp.h
@@ -35,7 +35,7 @@ namespace KUserTimestamp
      * if 0 is given. Do not use unless you're really sure what you're doing.
      * Consult focus stealing prevention section in kdebase/kwin/README.
      */
-    KINTERPROCESSWINDOWING_EXPORT void updateUserTimestamp(int time = 0);
+    KINTERPROCESSWINDOWING_EXPORT void updateUserTimestamp(unsigned long time = 0);
 };
 
 #endif
